Moving effectbuffer application to the frontend
