+++ gallifrey/core/src/main/java/gallifrey/core/CRDTEffect.java
@@ -1,11 +1,12 @@
-package gallifrey.backend;
+package gallifrey.core;
 
 import gallifrey.core.CRDT;
+import gallifrey.core.Effect;
 import gallifrey.core.VectorClock;
 
 public class CRDTEffect extends Effect {
     private static final long serialVersionUID = 10L;
-    CRDT crdt;
+    public CRDT crdt;
 
     public CRDTEffect(CRDT crdt, VectorClock time) {
         super(time);
