+++ gallifrey/core/src/main/java/gallifrey/core/GenericEffect.java
@@ -1,4 +1,4 @@
-package gallifrey.backend;
+package gallifrey.core;
 
 import gallifrey.core.GenericFunction;
 import gallifrey.core.VectorClock;
