+++ gallifrey/core/src/main/java/gallifrey/core/RMIInterface.java
@@ -7,8 +7,8 @@ import eu.antidotedb.client.GenericKey;
 
 public interface RMIInterface extends Remote {
 
-    public Object rmiOperation(GenericKey k, GenericFunction f) throws RemoteException, BackendRequiresFlushException;
+    public Snapshot rmiOperation(GenericKey k) throws RemoteException, BackendRequiresFlushException;
 
-    public Object rmiOperation(GenericKey k, GenericFunction f, VectorClock v) throws RemoteException;
+    public Snapshot rmiOperation(GenericKey k, VectorClock v) throws RemoteException;
 
 }
