+++ gallifrey/core/src/main/java/gallifrey/core/Snapshot.java
@@ -1,4 +1,4 @@
-package gallifrey.backend;
+package gallifrey.core;
 
 import java.io.Serializable;
